refactor(test): rename mock.ts to messaging.ts and update imports (#580)
